refactor: move RichSet implementation to richset/_richset.py
extract the 557-line class body from __init__.py into a dedicated
private module; __init__.py becomes a thin re-export wrapper

this separates the implementation detail (_richset.py) from the
public package entry point (__init__.py), making it easier to
split sections further in follow-up work
